microservices/object-remover-service/requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Image processing
opencv-python>=4.8.1.78
Pillow>=10.1.0
numpy>=1.24.3

# AI/ML dependencies
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
diffusers>=0.21.4
transformers>=4.35.2
accelerate>=0.24.1

# Cloud storage
cloudinary>=1.36.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0